chore: extracted snapshot logic
