+++ crates/dry_run_core/src/schema/mod.rs
@@ -2,6 +2,7 @@ pub mod bloat;
 mod hash;
 mod introspect;
 pub mod profile;
+mod snapshot;
 mod types;
 pub mod vacuum;
 
@@ -10,4 +11,5 @@ pub use introspect::{
     fetch_is_standby, introspect_activity_stats, introspect_planner_stats, introspect_schema,
 };
 pub use profile::*;
+pub use snapshot::*;
 pub use types::*;
