+++ crates/dry_run_core/src/schema/snapshot_tests.rs
@@ -1,3 +1,4 @@
+use super::super::types::*;
 use super::*;
 
 #[test]
